second work - petsitter

+++ src/com/jcpdev/dto/PetProfile.java
@@ -13,7 +13,7 @@ public class PetProfile {
 	private String pet;			// 견종		
 	private String p_name;		// 이름	
 	private String p_gender;	// 성별
-	private int p_weight;	// 몸무게
+	private double p_weight;	// 몸무게
 	private String p_birth; 	// 생년월일
 	private String p_neu; 		// 중성화여부
 }
